% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_funs.R
\name{api_gettags}
\alias{api_gettags}
\title{get tags of a video}
\usage{
api_gettags(aid = 3653764, sleep = 1)
}
\arguments{
\item{aid}{aid ov the video}
}
\value{
list
}
\description{
get tags of a video
}
\examples{
# https://www.bilibili.com/video/av2441580/
api_gettags(aid = 2441580)
}
