% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_funs.R
\name{api_getuploads_fp}
\alias{api_getuploads_fp}
\title{api_getuploads}
\usage{
api_getuploads_fp(mid, kw = "")
}
\arguments{
\item{mid}{up's member id}

\item{kw}{keyword for filtering results}
}
\value{
a dataframe of all uploaded videos of mid
}
\description{
api_getuploads
}
\examples{
api_getuploads_fp(5382023)
}
