% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_funs.R
\name{api_is_aid_alive}
\alias{api_is_aid_alive}
\title{test if a video is alive}
\usage{
api_is_aid_alive(aid = 56447372, sleep = 1)
}
\arguments{
\item{aid}{one aid at a time.}

\item{sleep}{interval of quest if multiple aid is supplied}
}
\value{
response code
}
\description{
test if a video is alive
}
\examples{
api_is_aid_alive(aid = 56447372, sleep = 1.0)

}
